generated_code/fixtures/tests/test_executeunittests_fixture.py
from code_autoeval.llm_model.utils.extraction.parse_unit_test_coverage import ParseUnitTestCoverage
from code_autoeval.llm_model.utils.preprocess_code_before_execution import PreProcessCodeBeforeExecution
from code_autoeval.llm_model.utils.extraction.parse_unit_test_coverage import ParseUnitTestCoverage
from code_autoeval.llm_model.utils.preprocess_code_before_execution import PreProcessCodeBeforeExecution
import pytest
from generated_code.fixtures.fixtures.executeunittests_fixture import fixture_mock_executeunittests
from code_autoeval.llm_model.utils.execute_unit_tests import ExecuteUnitTests
from pathlib import Path
from typing import Dict
import subprocess
def test_mock_executeunittests(fixture_mock_executeunittests):
    assert isinstance(fixture_mock_executeunittests, ExecuteUnitTests)
    assert hasattr(fixture_mock_executeunittests, 'coverage_result')
    assert fixture_mock_executeunittests.coverage_result is None
    assert hasattr(fixture_mock_executeunittests, 'file_path')
    assert isinstance(fixture_mock_executeunittests.file_path, Path)
    assert hasattr(fixture_mock_executeunittests, 'test_file_path')
    assert isinstance(fixture_mock_executeunittests.test_file_path, Path)
    assert hasattr(fixture_mock_executeunittests, 'absolute_path_from_root')
    assert isinstance(fixture_mock_executeunittests.absolute_path_from_root, Path)
    assert hasattr(fixture_mock_executeunittests, 'unique_imports_dict')
    assert isinstance(fixture_mock_executeunittests.unique_imports_dict, Dict)
    assert isinstance(fixture_mock_executeunittests, PreProcessCodeBeforeExecution)
    assert isinstance(fixture_mock_executeunittests, ParseUnitTestCoverage)